refactor: move analysis + plots into pdft_benchmarks package
- src/pdft_benchmarks/analysis.py: was top-level analyze.py.
- src/pdft_benchmarks/plots/loss_trajectories.py + rate_distortion.py:
  were top-level plots/. plots/__init__.py now re-exports both
  plot_loss_trajectories + plot_rate_distortion.
- src/pdft_benchmarks/_report.py: was top-level generate_report.py.
- experiments/post_run_analysis.py: was top-level (it's a CLI driver
  for the post-hoc analysis; lives with the other runners).

Tests adjusted accordingly (test_analyze.py -> test_analysis.py;
imports of `from analyze`, `from plots.X`, `from generate_report`
all rewritten).

Final repo layout:
  src/pdft_benchmarks/{__init__, presets, datasets/, bases, baselines,
                       evaluation, reporting, pipeline, analysis,
                       _training, _report, plots/}
  experiments/{div2k_8q_circuit_vs_classical, div2k_8q_blocked_vs_blockdct,
               div2k_10q_circuit, quickdraw, post_run_analysis}.py
  tests/  (mirrors src/ — datasets/, ...)

No top-level Python files remain (was 14 .py + 3 .sh).
