+++ experiments/post_run_analysis.py
@@ -43,7 +43,7 @@ import pdft
 
 import jax.numpy as jnp  # noqa: E402  - imported after pdft per CLAUDE.md §5
 
-from analyze import analyze_reconstructions  # noqa: E402
+from pdft_benchmarks.analysis import analyze_reconstructions  # noqa: E402
 from pdft_benchmarks.baselines import (  # noqa: E402
     block_dct_compress,
     block_fft_compress,
