--- plots/__init__.py
@@ -1,1 +1,0 @@
-"""Plot module dir. Imported as a package via benchmarks/_bootstrap.py."""
