+++ src/pdft_benchmarks/plots/__init__.py
@@ -1,0 +1,6 @@
+"""Plot helpers for benchmark results."""
+
+from .loss_trajectories import plot_loss_trajectories
+from .rate_distortion import plot_rate_distortion
+
+__all__ = ["plot_loss_trajectories", "plot_rate_distortion"]
