+++ src/pdft_benchmarks/plots/loss_trajectories.py
